Security Options box.
I did not include a way to check the CICQEventTag and close the window
or notify the user that it was succesfull.  I'm preparing to do an overhaul
of the way that my plugin uses the CICQEventTag and will hopefully have that
committed over the weekend.  Just use the Network Log window to see if it
worked or not until then.