Better chats, still in progress though, so don't report any bugs, as this is going to be changed still, and fixed of course!
Options window, not working really, that's still in progress too.
It defaults to not showing users on your ignore list on your contact list.
To see them.. open the options window, click on "Show Ignore Users" and then
hit the close button that your window manager gives you (Not the "Done button")
and then refresh your list from the System menu.

Removed file stuff that has been changed.. need to work on that as well.