Added: void ICQUser::SetAutoSecure(bool) bool ICQUser::AutoSecure()
Uses a new flag for the m_nAutoAccept (AUTO_SECURE)

Plugins will have to use this in the contact_list_refresh functions.  The
easiest way I thought would work is with a list of uins that use it.  Look
at contact_list.cpp for how I did it if you're curious.