Extending the OnEvents to allow this to be handled by the plugin to allow better sound integration with the different desktops (KDE/GNOME).
To get the new signals (SIGNAL_ONEVENT) you set
COnEventManager->SetCommandType(ON_EVENT_BY_PLUGIN).

signal->Signal() == SIGNAL_ONEVENT for this new signal.
The SubSignal() is containing the eventnumber as defined in licq_onevent.h

The class CICQSignal now have a new attribute Parameters(), this is to allow
passing the parameters for the OnEvent command. If this parameter wouldn't
be there it would probably require the plugin to duplicate some work from the
daemon.