Initial server side contact list support.
The way it works is:

Every time you log on it will retrieve the server list.
It will add groups that the server list has if they are not local.
Contacts are added from the server list.
If they are already on the list they will be be in their group that the server s
ays they are in, as well as the local groups they are in.
If they are not already on the list they will  only be in the group that the ser
ver says they are in.
Visibile and Invisible contacts are put in the appropriate list according to wha
t the server says.

It needs to update new and removed users.
It needs to update new, removed, and renamed groups.

Once again this is preliminary.