- Removed 'Translations' from the Options screen. The per-user Encodings are both a more flexible (you can talk in different languages with different people; you don't rely on the system locale) and sensible (in sense of Qt using Unicode internally) way to support multiple encodings.
*** IF ANYONE HAS COMPLAINTS ABOTU THIS CHANGE, please mail me instead
of rolling it back. I've been using Encodings fine for the past half a year
and if something stops you from using them, I'd be happy to help. ***

- Instead of a global translation, the user can select a Default Encoding now
in Options.
This is the encoding which'll be set for newly added users.

- Now we show less encodings by default (specifically, only encodings
used on Windows + UTF-8). This would be less confusing for most users
(huh? 3 Russian encodings? Which is the right one?). To enable a full
list of encodings, check the 'Show all encodings' option in the Options
dialog.