Fixing a bug which caused offline messages to be duplicated in the MsgView. What caused it is that they were added both in the ctor (by polling for new messages) and in UserUpdated.
Polling is required for cases when the messages were received (and the
appropriate signal was already delivered) before this dialog popped up.
Receiving signals is required so that we'd get messages received
during the existance of the dialog dynamically added to the dialog.

Thus we cannot give up either. The solution is to add a member which
holds the Id of the last processed message event. Since Ids are
incremental, we can compare the Id of the new event and determine
whether we processed it (and possibly the events following it)
already.