Don't trim the password anymore since spaces can be part of the password at the end or at the beginning.
If this is a problem for you then you must edit your owner.uin file to be like this:
Password = <password>

Please note that there is one space after the equal sign.  Anything after that space is considered the password.  This will only be a problem if you hand modified your password in owner.uin and did not follow this format.  Which shouldn't be much of a problem anymore since of the modifying local password dialog is in the qt-gui plugin.