* Let the protocol plugins delete the signal, not the daemon.  Reverts to origi nal behaviour
* Integrate the GPG patch by Richard Hirner, but slightly modified.  Works with protocol plugins.  Each protocol plugin will have to call the Encrypt() function, as this simplifies the Encryption flag and displaying of the unencrypted message to the message window.

To use the GPG feature please check the licq_gpg.conf file for information.
It is best to have gpg-agent running.  This will allow you to not be forced to
save your gpg password in plaintext to the config file.  Now we can send encryp
ted messages through the server, and offline encrypted messages!