WARNING: This can break some behavior of the plugins.
An ID of a ICQ Protocol user will not be stripped of all spaces and converted to lower case before it is stored in the user hash.  This means if a plugin uses strcmp to make sure of a particular user, it will have to be updated to use ICQUser::MakeRealId().  This function converts a user Id to the stored format.  I wrote the fix for the qt-gui, just take a look at that to fix other plugins (if they are broken from this).  This will need to be completely checked before 1.3.0 is released.

Why was this done? AIM screen names do not care about case and spaces.  This fixes a bug if you add an AIM user's name differently than how they created it.  It would be added, but Licq would receive the real screen name and not be able to find that user in the hash.