Disable the Main Window typing notification for other protocols.
MSN would require a timer for each user that types, since it doesn't notify when the user has stopped typing. This would add a large amount of overhead to the Qt-GUI, which is not worth it. Perhaps a different method can be done (i.e. Automatically open a window from the user) in the future.