Don't crash on drag&drop if there already is a chat window opened. Closes #1368.
The bug was that the pointer to the existing chat window was casted (static_cast) to a specific class (e.g. the send file dialog class). But if the chat window was already open the pointer could be pointing to a sibling class (e.g. the send message dialog class) and this resulted in a crash.

Now we use dynamic_cast instead, and if it fails, we try to change the dialog type to match that that was dragged&dropped.