Redo the "only one instance of Licq" logic.
Instead of checking if the pid in licq.pid is running, use advisory locking. First we grab a write lock on licq.pid. If it fails, this means that licq is already running, otherwise we're alone. By never closing the file, we keep the lock until licq is killed (normally or abnormally) when the kernel will close all open files and remove all locks.

Fixes #1353.