My commit in r4677 (and r4678) broke po installation if configured was generated with autoconf 2.60. So instead of going backwards, we'll use Makefile.in.in revision 1.25 from gettext cvs. But with one difference: {{{ localedir = $(datadir)/locale }}} instead of {{{ localedir = @localedir@ }}}
This should work with both autoconf 2.59 and 2.60. I've tested it on both Debian Sarge (autoconf 2.59) and Debian Sid (2.60) and it works for me (tm).But if someone else could try it as well, that would be great! Do make install DESTDIR=/tmp/licq-test if you just want to test the installation without messing with your current installation.