Patch from Anders Olofsson to make the history layout configurable. Also adds a new style format that can be used in the history. Fixes part of #1449.
With this commit, <font> tags are removed before the message is displayed. This means that the sender can no longer choose the font or color to show the message in. IMHO, this is a good thing. But if you have an other opinion, speak up and maybe I'll make it configurable.