Revised the user id to be a string consisting of protocol id and account id Compared to the previous int, this user id string has the advantage of being usable for users not in the list, i.e. no separate handling is needed for search and add functions. The id is also guaranteed to be unique even if users are removed and added to the contact list.
A typedef has also been added to define the user id. This allows future changes of the user id type to only require minimal changes to the existing code. It also makes it a more distinct data type in variable lists.

A new header file (licq_types.h) has been created to hold type definitions (such as the user id). This allows headers that need it to only include the bare minimum of definitions instead of being dependant on the entire licq_user.h.