Reverted r7881 and fixed the crash properly
When a KApplication is started a new KMessageBoxMessageHandler (from libkdeui)
is created and is installed as the message handler. The call to
setMessageHandler() triggers KMessage to create a static object that will
delete the handler when libkdecore is unloaded.

The problem is that when kde4-gui is unloaded (which happens before Licq exits)
only libkdeui is unloaded; something makes libkdecore stay loaded. When Licq
exits, all libs are unloaded. Before libkdecore is unloaded, the static objects
are deleted. In the destructor of the static object created by KMessage a call
to the destructor for KMessageBoxMessageHandler is attempted. But that code is
in libkdeui which already has been unloaded, resulting in a crash.