Added a filter for user events
The filter can match events based on protocol, event type or message contents.
Filter rules can be disabled/enabled without removing them from the
configuration. Message contents is matched with a regular expression.
A match can allow the event, just drop it, or log the event as normal without
notifying the user.
Only events from unknown contacts are filtered, contacts in the list are
always trusted. If no rule applies, the default is to accept the event.

The default configuration is set to drop a few spam messages that I have seen
repeatedly the last months. It will also silence (but log) any authorization
request containing a URL as those are generally also spam.

Limitations:
- Currently there is no GUI, only the daemon logic and a configuration file.
- The filter does not understand non-ascii characters making the rules for
  cryllic characters very unreadable.
- Counters should be added to let the user see which rules get matches.