Update graphical plugins for new plugin base
For Qt4-Gui and Aosd, new plugin classes are added rather than trying
to make an existing classes inherit Licq::GeneralPlugin.

For Osd a class was created to hold the essential functions needed to
inherit Licq::GeneralPlugin. It is left as a future improvement to
change the remaining functions to also be members of the OsdPlugin
class.