api: Change plugin bases to be object based
The new plugin base exports only a single symbol which is a structure
containing everything Licq need to know to load the plugin. Currently
the contents is a magic value, the Licq version the plugin is built and
pointers to a functions for creating and deleting the plugins.

Plugin loading is only allowed if the magic value is correct and the
Licq version matches. This should prevent crashes later if plugins are
built for a Licq version with a different api. It is assumed that the
last digit in the version is only used for minor releases that doesn't
change the api and plugin loading will be allowed between for example
1.6.0 and 1.6.2, but not between 1.6.0 and 1.7.0.

The factory function must create and return a plugin object derived
from either Licq::GeneralPlugin or Licq::ProtocolPlugin. This object
is then the interface for communicating between Licq and the plugin.

This means that future additions to the plugin interface can be done
without having to export additional symbols for each plugin.
It should also allow plugins to communicate more directly with each
other by sharing a header file and casting the other plugins pointer
to the known subclass.

Additionally:
- Extra comments have been added in plugin interface headers to clarify
  which functions are to be used by the plugin itself and which are
  used by Licq and other plugins.
- The PluginManager uses a custom deleter for the plugins that makes
  sure the plugin instance is completely destroyed before deleting
  the DynamicLibrary instance that will close the library file.
- Plugin base classes contains implementations of configFile() and
  isEnabled() allowing plugins to skip overriding.
- Plugins no longer need to register and unregister with the
  PluginManager, instead they can get the read pipe and set their
  signal mask through the plugin base class.