Drop old unmaintained conversion scripts
Server side list is standard for all protocols and I could not find
that any of this scripts migrated anything additional (like history).
As none of these scripts have been updated for multi protocol support,
which was introduced with Licq 1.3.0 (7 years ago), it should
be safe to assume that noone needs these anymore.