api: Don't define entire derived classes in header
With all functions defined in the header, the compiler will not make a
single typeinfo for the class. This caused dynamic_cast in plugins to
fail sometimes as the object was created with a pointer to the daemons
typeinfo data but dynamic_cast was called with a different typeinfo
built into the plugin.
The key seems to be to have at least one virtual member that is
overriden in every subclasses, e.g. the virtual destructors are enough.