Remove console plugin
As "discussed" on the licq-dev mailing list, the console plugin is in
bad shape and noone is willing to maintain it any longer.
As of now, it is no longer maintained as a part of Licq.

Rather than keeping dead code around, I've removed all the files. If
anyone would ever like to revive it, the code can be brought back from
an older commit.