licq: Move user data files into subdirs per owner
User files for each owner is placed in users/<owner>.<ppid>/ so files
will be unique even if multiple owners exist with the same user added.

Note: The condition for performing the upgrade is the Licq version
number in licq.conf. As the dev version in git master is already set to
1.8.0, people already running with this version must edit licq.conf
manually to a lower value (for example "1079") before running Licq to
trigger the upgrade to be performed.