core: implement new api for handling plugins
Plugins implement one of the interfaces GeneralPluginInterface or
ProtocolPluginInterface depending on type. The instance implementing this is
separate from the instance of Licq::GeneralPlugin or Licq::ProtocolPlugin that
provides the interface to the plugin for other plugins.

Both Licq::GeneralPlugin and Licq::ProtocolPlugin are extended by
LicqDaemon::GeneralPlugin and LicqDaemon::ProtocolPlugin respectively to
provide an extended interface for internal use in plugin manager.

This provides a cleaner separation between plugins and the core.