qt4-gui: use Qt signals instead of plugin helper to handle signals/events
If a Qt signal is sent to a object that lives in another thread, the signal is
queued in that object's event loop and delivered in that thread.

Since this was what the helper class did, but using a std::queue and a pipe, it
could be removed in favour of the native solution.