msn: Handle typing status timeouts
The MSN protocol does not send any notification when the peer stops
typing. A timer is added to the protocol code to clear the typing
status after 10 seconds from the last received typing notification.
We also need to resend the our typing notifications every 5 seconds
to keep it activate.
This logic was previously done in the GUI.