dbus: New plugin providing a D-Bus interface
This plugin will allow Licq to be remotely controlled from other
applications via D-Bus. It also emits signals for events allowing osd
style notification applications to be created without building new Licq
plugins.

The current API is limited to just a few simple operations but should
be easy to extend depending on user needs.

The D-Bus access is separated from the Licq plugin logic to make it
reuseable. (For example to make a protocol plugin for the Skype API.)